requirements.txt
absl-py==0.13.0
airsim==1.6.0
argon2-cffi==21.1.0
arm-mango==1.2.0
astor==0.8.1
async-generator==1.10
attrdict==2.0.1
attrs==21.2.0
backcall==0.2.0
baselines==0.1.6
bleach==4.1.0
cached-property==1.5.2
certifi==2022.5.18.1
cffi==1.14.6
charset-normalizer==2.0.12
click==8.0.1
cloudpickle==1.6.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
filterpy==1.4.5
gast==0.5.2
google-pasta==0.2.0
grpcio==1.40.0
gym==0.19.0
h5py==2.10.0
idna==3.3
imageio==2.13.5
importlib-metadata==4.8.1
install==1.3.4
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.4
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mpmath==1.2.1
msgpack==1.0.2
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==5.2.0
numba==0.53.1
numpy==1.19.5
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
ptyprocess==0.7.0
pyaml==21.10.1
pybullet==3.1.9
pycparser==2.20
pyglet==1.5.0
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.11.0
requests==2.27.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
stable-baselines3==1.2.0
sympy==1.8
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
tornado==4.5.2
tqdm==4.62.2
traitlets==4.3.3
typing-extensions==3.10.0.2
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.5.0
